Move ExternalResources to utils module (like silx)
