+++ fabio/utils/ExternalResources.py
